+++ LogicBuilder.App.Spa.Business.Tests/Requests/DetailRequestTest.cs
@@ -2,6 +2,8 @@
 {
     using LogicBuilder.App.Spa.Business.Requests;
     using LogicBuilder.App.Spa.Business.ScreenSettings.Views;
+    using LogicBuilder.Domain;
+    using System.Collections.Generic;
     using System.Text.Json;
 
     public class DetailRequestTest
@@ -11,7 +13,8 @@
         {
             var model = new DetailRequest
             {
-                Entity = null,
+                PersistentFlowItems = new Dictionary<string, object> { ["UserId"] = 1, ["UserName"] = "Smith101", ["UserRating"] = 9.5 },
+                Entity = new TestModel { Id = 101, FirstName = "John" },
                 ViewType = ViewType.Detail,
                 CommandButtonRequest = new CommandButtonRequest
                 {
@@ -21,13 +24,24 @@
             };
 
             var json = JsonSerializer.Serialize(model);
-            var result = JsonSerializer.Deserialize<DetailRequest>(json);
+            var result = JsonSerializer.Deserialize<DetailRequest>(json, SerializationOptions.Default);
 
             Assert.NotNull(result);
+            Assert.NotNull(result.PersistentFlowItems);
+            Assert.Equal((int)model.PersistentFlowItems["UserId"], (int)result.PersistentFlowItems["UserId"]);
+            Assert.Equal((string)model.PersistentFlowItems["UserName"], (string)result.PersistentFlowItems["UserName"]);
+            Assert.Equal((double)model.PersistentFlowItems["UserRating"], (double)result.PersistentFlowItems["UserRating"]);
             Assert.Equal(model.ViewType, result.ViewType);
-            Assert.Null(result.Entity);
+            TestModel resultEntity = Assert.IsType<TestModel>(result.Entity, exactMatch: false);
+            Assert.Equal(((TestModel)model.Entity).Id, resultEntity.Id);
             Assert.NotNull(result.CommandButtonRequest);
             Assert.Equal(model.CommandButtonRequest.NewSelection, result.CommandButtonRequest.NewSelection);
+        }
+
+        private class TestModel : BaseModel
+        {
+            public int Id { get; set; }
+            public string FirstName { get; set; } = "";
         }
     }
 }