+++ LogicBuilder.App.Spa.Business.Tests/ScreenSettings/FlowSettingsTest.cs
@@ -3,6 +3,7 @@
     using LogicBuilder.App.Spa.Business.ScreenSettings;
     using LogicBuilder.App.Spa.Business.ScreenSettings.Navigation;
     using LogicBuilder.App.Spa.Business.ScreenSettings.Views;
+    using System.Collections.Generic;
     using System.Text.Json;
 
     public class FlowSettingsTest
@@ -12,6 +13,7 @@
         {
             var model = new FlowSettings
             (
+                new Dictionary<string, object> { ["UserId"] = 1, ["UserName"] = "Smith101", ["UserRating"] = 9.5 },
                 new FlowState
                 {
                     Driver = "Driver",
@@ -24,13 +26,17 @@
                     BrandText = "Brand",
                     CurrentModule = 2
                 },
-                new LogicBuilder.App.Spa.Business.ScreenSettings.Views.ScreenSettings<string>("dialog-settings", [], ViewType.Detail)
+                new ScreenSettings<string>("dialog-settings", [], ViewType.Detail)
             );
 
             var json = JsonSerializer.Serialize(model);
-            var result = JsonSerializer.Deserialize<FlowSettings>(json);
+            var result = JsonSerializer.Deserialize<FlowSettings>(json, SerializationOptions.Default);
 
             Assert.NotNull(result);
+            Assert.NotNull(result.PersistentFlowItems);
+            Assert.Equal((int)model.PersistentFlowItems["UserId"], (int)result.PersistentFlowItems["UserId"]);
+            Assert.Equal((string)model.PersistentFlowItems["UserName"], (string)result.PersistentFlowItems["UserName"]);
+            Assert.Equal((double)model.PersistentFlowItems["UserRating"], (double)result.PersistentFlowItems["UserRating"]);
             Assert.NotNull(result.FlowState);
             Assert.Equal(model.FlowState!.Driver, result.FlowState.Driver);
             Assert.NotNull(result.NavigationBar);